Added reactions and experience. Also added auto profanity checker. Reactions is commented out until I figure out a better way to implement it though

+++ lib/cogs/welcome.py
@@ -1,4 +1,4 @@
-from discord.ext.commands import command, has_permissions, Cog, CheckFailure, TextChannelConverter
+from discord.ext.commands import command, has_permissions, Cog, CheckFailure, TextChannelConverter, MissingPermissions
 from ..db import db
 
 class Welcome(Cog):
@@ -27,20 +27,24 @@ class Welcome(Cog):
     
     @welcome_message.error
     async def welcome_message_error(self, ctx, exception):
-        if isinstance(exception, CheckFailure):
-            await ctx.send("You need the Manage Messages permissions to do that.")
+        if isinstance(exception, MissingPermissions):
+            await ctx.send("User does not have permissions to manage server.")
 
     @Cog.listener()
     async def on_memeber_join(self, member):
         send_message = db.field("SELECT WelcomeMessage FROM guilds WHERE GuildID =?", self.ctx.guild.id)
         if send_message == "enabled":
             channel = db.field("SELECT WelcomeChannelID FROM guilds WHERE GuildID =?", self.ctx.guild.id)
-            db.execute("INSERT INTO exp (UserID) VALUES (?)", member.id)
             await self.bot.get_channel(channel).send(f"Welcome {member.mention}! Please remember to adhere to the rules and have fun!")
+        exp = db.field("SELECT Experience FROM guilds WHERE GuildID =?", self.ctx.guild.id)
+        if exp == "enabled":
+            db.execute("INSERT INTO exp (UserID) VALUES (?)", member.id)
 
     @Cog.listener()
     async def on_member_remove(self, member):
-        db.execute("DELETE FROM exp WHERE UserID = ?", member.id)
+        exp = db.field("SELECT Experience FROM guilds WHERE GuildID =?", self.ctx.guild.id)
+        if exp == "enabled":
+            db.execute("DELETE FROM exp WHERE UserID = ?", member.id)
     
 def setup(bot):
     bot.add_cog(Welcome(bot))
